Give the gutter a Float and the archive preview its entry type
Two compile errors in :app's main source set, both introduced by the previous
commit and both caught by CI rather than locally:

- `lineTops` wrote `layout.size.height` — an `IntSize`'s `Int` — into a
  `FloatArray`. The last entry is the layout's own bottom edge, so the widening
  is the whole fix.
- `ArchiveEntryPreviewActivity` names `ArchiveEntry` in its `intent` helper
  without importing it. The name reads as if it were local because the file
  would live in that package if it were not under `ui.archive`, which is how it
  went unnoticed.

Neither changes behaviour: the first converts a value that was already
computed, the second resolves a name that was always meant.

+++ app/src/main/java/dev/eclipse/ssh/ui/editor/TextEditorScreen.kt
@@ -1283,7 +1283,7 @@ private fun lineTops(text: String, layout: TextLayoutResult): FloatArray {
         val offset = starts[position].coerceIn(0, known)
         tops[position] = layout.getLineTop(layout.getLineForOffset(offset))
     }
-    tops[starts.size] = layout.size.height
+    tops[starts.size] = layout.size.height.toFloat()
     return tops
 }
